apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalpzero-enterprise-api"
version = "0.1.0"
description = "Canonical FastAPI backend for KalpZero Enterprise"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.12,<1.0.0",
  "PyYAML>=6.0.3,<7.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "pymongo>=4.13.2,<5.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "redis>=6.4.0,<7.0.0",
  "psycopg[binary]>=3.2.6,<4.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
  "beanie>=1.21.0,<2.0.0",
  "motor>=3.3.0,<4.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.5,<9.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
